PR is now dependent on https://github.com/functional-streams-for-scala/fs2/pull/1257, will have to rerun the build when that one is merged

+++ core/src/main/scala/org/http4s/StaticFile.scala
@@ -190,7 +190,7 @@ object StaticFile {
       end: Long,
       blockingExecutionContext: ExecutionContext
   ): EntityBody[F] =
-    readRange[F](f.toPath, blockingExecutionContext, DefaultBufferSize, start.toInt, end.toInt)
+    readRange[F](f.toPath, blockingExecutionContext, DefaultBufferSize, start, end)
 
   private def nameToContentType(name: String): Option[`Content-Type`] =
     name.lastIndexOf('.') match {
